pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["Our World In Data <tech@ourworldindata.org>"]

[tool.poetry.scripts]
etl = 'etl.command:main_cli'
reindex = 'etl.reindex:reindex_cli'
publish = 'etl.publish:publish_cli'
prune = 'etl.prune:prune_cli'
harmonize = 'etl.harmonize:harmonize'
backport = 'backport.backport:backport_cli'
backport-migrate = 'backport.migrate.migrate:cli'
bulk_backport = 'backport.bulk_backport:bulk_backport'
backport-datasync = 'backport.datasync.datasync:cli'
walkthrough = 'walkthrough.cli:cli'
fasttrack = 'fasttrack.cli:cli'
run_python_step = 'etl.run_python_step:main'
compare = 'etl.compare:cli'
generate_graph = 'etl.to_graphviz:to_graphviz'
version_tracker = 'etl.helpers:run_version_tracker_checks'
etl-match-variables = 'etl.match_variables:main_cli'
etl-translate-varmap = 'etl.variable_mapping_translate:main_cli'
etl-chart-suggester = 'etl.chart_revision.cli:main_cli'
etl-metadata-export = 'etl.metadata_export:cli'
etl-datadiff = 'etl.datadiff:cli'
etl-chartgpt = 'etl.chart_revision.v2.chartgpt:cli'

[tool.poetry.dependencies]
python = "^3.10"
owid-catalog = {path = "lib/catalog", develop = true}
walden = {path = "lib/walden", develop = true}
click = ">=8.0.1"
PyYAML = ">=5.4.1"
papermill = ">=2.3.3"
jupyterlab = ">=3.1.13"
# 2.2.0 was raising weird error, pin to 2.1.1 which is stable
mysqlclient = "2.1.1"
SPARQLWrapper = ">=1.8.5"
openpyxl = ">=3.0.9"
sh = "1.14.3"
Unidecode = ">=1.3.2"
python-dotenv = ">=0.19.0"
# we're not yet ready for frictionless 5.x.x as it raises an error for data://open_numbers/open_numbers/latest/bp__energy
frictionless = {version = "^4.40.8", extras = ["pandas"]}
regex = ">=2022.1.18"
wikipedia = ">=1.4.0"
# pandas updates can be dangerous, it is better to pin the version and carefully update it manually
pandas = "1.5.2"
numpy = ">=1.22.1"
pydantic = ">=1.9.0"
structlog = ">=21.5.0"
sqlmodel = "^0.0.6"
rich = ">=12.1.0"
owid-datautils = {path = "lib/datautils", develop = true}
rich-click = ">=1.5.1"
tenacity = ">=8.0.1"
simplejson = ">=3.17.6"
bugsnag = ">=4.2.1"
xlrd = ">=2.0.1"
PyPDF2 = ">=2.11.1"
dvc = {extras = ["s3"], version = ">=2.56.0"}
"ruamel.yaml" = ">=0.17.21"
owid-repack = {path = "lib/repack", develop = true}
mkdocs-jupyter = "^0.22.0"
gitpython = "^3.1.30"
rapidfuzz = "^2.13.7"
streamlit = "^1.22.0"
# pinned because of frictionless, remove this pin when frictionless is updated
fsspec = "2022.11.0"
openai = "^0.27.7"
pdfplumber = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.1.2"
flake8 = ">=3.9.2"
watchdog = ">=2.1.5"
PyYAML = ">=5.4.1"
argh = ">=0.26.2"
jupyter = ">=1.0.0"
ipython = ">=7.27.0"
matplotlib = ">=3.4.3"
jupytext = "^1.14.6"
jupyter_nbextensions_configurator = ">=0.4.1"
plotly = ">=5.7.0"
cookiecutter = ">=2.1.1"
isort = ">=5.12.0"
pyright = "1.1.288"
types-PyYAML = ">=6.0.12"
hydra-core = ">=1.3.0"
mkdocs-material = "<9.0.0"
pywebio = ">=1.8.2"
black = {version = "22.3.0", extras = ["jupyter"]}
boto3-stubs = { extras = ["s3"], version = ">=1.21.16" }
# unpinning pandas-stubs would introduce tons of type errors
pandas-stubs = "1.2.0.62"
notebook = {version = "^7.0.0a1", allow-prereleases = true}

[tool.black]
exclude = ".ipynb_checkpoints|walkthrough/.*_cookiecutter|.*playground.*.ipynb"
line-length = 120

[tool.isort]
profile = "black"
extend_skip = [".ipynb_checkpoints", "walkthrough/meadow_cookiecutter"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
exclude = [
    "walkthrough/*_cookiecutter",
    "**/node_modules",
    "**/__pycache__",
    ".git",
    "**/.venv",
    "**/.ipynb_checkpoints",
]
